+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 ## [Unreleased]
 ### Fixed
 * StackPanel, UniformPanel and ItemsPanel now take care about items be set on collapsed
+* The UniformPanel or Uniform ItemsPanel does not crash anymore if the height or width of the children became negative
 
 ## [1.8.0] - 2021-08-05
 ### Added